+++ CHANGELOG.md
@@ -38,15 +38,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - Added `state` field.
     - Added `scancode` field.
     - Added `keycode` field.
-  - Added `MouseMoved` variant.
+  - Added `MouseMove` variant.
     - Added `x` field.
     - Added `y` field.
-  - Added `RawMouseMoved` variant.
+  - Added `RawMouseMove` variant.
     - Added `delta_x` field.
     - Added `delta_y` field.
   - Added `MouseButton` variant.
     - Added `state` field.
     - Added `button` field.
+  - Added `MouseScroll` variant.
+    - Added `delta_x` field.
+    - Added `delta_y` field.
 - Added `ToInput` trait.
   - Added `to_input()` method.
 - Added `keyboard` module.